requirements.txt
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
beautifulsoup4==4.12.3
yfinance==0.2.52
pyyaml==6.0.2
pytz==2024.2
pandas==2.2.3  # data wrangling for prices/signals
numpy==2.1.3   # numeric helpers
requests==2.32.3  # HTTP for Finviz scrapes
beautifulsoup4==4.12.3  # parse Finviz HTML
yfinance==0.2.52  # Yahoo Finance OHLCV download
pyyaml==6.0.2  # read settings.yaml
pytz==2024.2  # timezone handling if needed
